# docs/swagger.yaml
basePath: /api
definitions:
...
  databasehandler.ConfigPayload:
    properties:
      auto_conversion:
        type: string
      create_preview:
        type: boolean
    type: object
  databasehandler.DatabaseCreatePayload:
    properties:
      config:
        $ref: '#/definitions/databasehandler.ConfigPayload'
      content_type:
        type: string
      custom_fields:
        items:
          $ref: '#/definitions/databasehandler.DatabaseCustomField'
        type: array
      housekeeping:
        $ref: '#/definitions/databasehandler.HousekeepingPayload'
      name:
        type: string
    type: object
  databasehandler.DatabaseCustomField:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
  databasehandler.DatabaseResponse:
    properties:
      config:
        $ref: '#/definitions/databasehandler.ConfigPayload'
      content_type:
        type: string
      custom_fields:
        items:
          $ref: '#/definitions/databasehandler.DatabaseCustomField'
        type: array
      housekeeping:
        $ref: '#/definitions/databasehandler.DatabaseResponseHK'
      name:
        type: string
      stats:
        $ref: '#/definitions/databasehandler.DatabaseResponseStats'
    type: object
  databasehandler.DatabaseResponseHK:
    properties:
      disk_space:
        description: e.g. "10G"
        type: string
      interval:
        description: e.g."10min"
        type: string
      max_age:
        description: e.g. "365d"
        type: string
    type: object
  databasehandler.DatabaseResponseStats:
    properties:
      entry_count:
        type: integer
      total_disk_space_bytes:
        type: integer
    type: object
  databasehandler.DatabaseUpdatePayload:
    properties:
      config:
        $ref: '#/definitions/databasehandler.ConfigPayload'
      housekeeping:
        $ref: '#/definitions/databasehandler.HousekeepingPayload'
    type: object
  databasehandler.HousekeepingPayload:
    properties:
      disk_space:
        type: string
      interval:
        type: string
      max_age:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: denglerchr@gmail.com
    name: Christian Dengler
    url: https://www.swcd.lu
  description: This is a server for a image, audio and file storage with integrated
    web-ui.
  title: SWCD MediaHub-API
  version: 2.0.0-beta1
paths:
...
  /database:
    post:
      consumes:
      - application/json
      description: Creates a new database with custom fields and a dedicated entry
        table.
      parameters:
      - description: Database Metadata
        in: body
        name: database
        required: true
        schema:
          $ref: '#/definitions/databasehandler.DatabaseCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Invalid request payload or missing name
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Database name already in use
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to create database or storage folder
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create a new database
      tags:
      - database
  /database/{dbname}:
    delete:
      description: Deletes a database, its entry table, and all of its associated
        entries and metadata.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Missing dbname path parameter or invalid name
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to delete database record or folder
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete a database
      tags:
      - database
    get:
      description: Retrieves details, custom fields, and statistics for a specific
        database.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Missing dbname path parameter
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to calculate stats
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get database details
      tags:
      - database
    put:
      consumes:
      - application/json
      description: Updates the housekeeping rules for a specific database.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Housekeeping Rules or Config flag
        in: body
        name: housekeeping
        required: true
        schema:
          $ref: '#/definitions/databasehandler.DatabaseUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Invalid request payload or missing dbname path parameter
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to update database
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update database housekeeping rules
      tags:
      - database
...
  /databases:
    get:
      description: Retrieves a list of all available databases and their statistics.
      produces:
      - application/json
      responses:
        "200":
          description: Returns an empty array if no databases exist
          schema:
            items:
              $ref: '#/definitions/databasehandler.DatabaseResponse'
            type: array
        "500":
          description: Failed to retrieve databases
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all databases
      tags:
      - database
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"